osint_platform/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
python-dotenv==1.0.0
python-json-logger==2.0.7

# OSINT Tools Integration
beautifulsoup4==4.12.3
lxml==5.1.0

# Database (optional)
sqlalchemy==2.0.25

# Security
cryptography==42.0.0

# Utils
colorama==0.4.6
rich==13.7.0
